add checksums for all data, update index format

+++ scripts/make_ikala_index.py
@@ -1,10 +1,11 @@
 import argparse
 import csv
-import os
 import glob
-import json
 import hashlib
+import json
+import os
 from urllib import request
+
 
 ID_MAPPING_URL = "http://mac.citi.sinica.edu.tw/ikala/id_mapping.txt"
 IKALA_INDEX_PATH = "../mir_dataset_loaders/indexes/ikala_index.json"
@@ -37,8 +38,8 @@ def make_ikala_index(ikala_data_path):
         [os.path.basename(f).split('.')[0] for f in lyrics_files])
 
     id_map_path = os.path.join(ikala_data_path, "id_mapping.txt")
-
-    request.urlretrieve(ID_MAPPING_URL, filename=id_map_path)
+    if not os.path.exists(id_map_path):
+        request.urlretrieve(ID_MAPPING_URL, filename=id_map_path)
 
     with open(id_map_path, 'r') as fhandle:
         reader = csv.reader(fhandle, delimiter='\t')
@@ -51,17 +52,32 @@ def make_ikala_index(ikala_data_path):
     ikala_index = {k: {} for k in track_ids}
     for key in ikala_index.keys():
         songid = key.split('_')[0]
-        ikala_index[key]['singer_id'] = singer_map[songid]
-        ikala_index[key]['song_id'] = key.split('_')[0]
-        ikala_index[key]['section'] = key.split('_')[1]
-        ikala_index[key]['audio_path'] = "iKala/Wavfile/{}.wav".format(key)
-        ikala_index[key]['pitch_path'] = "iKala/PitchLabel/{}.pv".format(key)
-        ikala_index[key]['lyrics_path'] = "iKala/Lyrics/{}.lab".format(key)
-
         audio_checksum = md5(os.path.join(
             ikala_data_path, "Wavfile/{}.wav".format(key)))
+        pitch_checksum = md5(os.path.join(
+            ikala_data_path, "PitchLabel/{}.pv".format(key)))
+        lyrics_checksum = md5(os.path.join(
+            ikala_data_path, "Lyrics/{}.lab".format(key)))
 
-        ikala_index[key]['audio_checksum'] = audio_checksum
+        ikala_index[key] = {
+            'data_files': {
+                'audio': {
+                    'path': "iKala/Wavfile/{}.wav".format(key),
+                    'checksum': audio_checksum
+                },
+                'pitch': {
+                    'path': "iKala/PitchLabel/{}.pv".format(key),
+                    'checksum': pitch_checksum
+                },
+                'lyrics': {
+                    'path': "iKala/Lyrics/{}.lab".format(key),
+                    'checksum': lyrics_checksum
+                }
+            },
+            'singer_id': singer_map[songid],
+            'song_id': key.split('_')[0],
+            'section': key.split('_')[1]
+        }
 
     with open(IKALA_INDEX_PATH, 'w') as fhandle:
         json.dump(ikala_index, fhandle, indent=2)
